requirements.txt
loguru
httpx[http2]
GitPython